Feat: Added animation to clock
