+++ analog_clock/lib/components/clock.dart
@@ -1,6 +1,9 @@
+import 'dart:async';
 import 'package:analog_clock/constants.dart';
 import 'package:analog_clock/size_config.dart';
 import 'package:flutter/material.dart';
+
+import 'clock_painter.dart';
 
 class Clock extends StatefulWidget {
   @override
@@ -9,6 +12,16 @@ class Clock extends StatefulWidget {
 
 class _ClockState extends State<Clock> {
   DateTime _dateTime = DateTime.now();
+
+  @override
+  void initState() {
+    super.initState();
+    Timer.periodic(Duration(seconds: 1), (timer) {
+      setState(() {
+        _dateTime = DateTime.now();
+      });
+    });
+  }
 
   @override
   Widget build(BuildContext context) {
@@ -28,61 +41,14 @@ class _ClockState extends State<Clock> {
               ),
             ],
           ),
-          child: CustomPaint(
-            painter: ClockPainter(context, _dateTime),
+          child: RotatedBox(
+            quarterTurns: 3,
+            child: CustomPaint(
+              painter: ClockPainter(context, _dateTime),
+            ),
           ),
         ),
       ),
     );
-  }
-}
-
-class ClockPainter extends CustomPainter {
-  final BuildContext context;
-  final DateTme dateTime;
-
-  ClockPainter(this.context, this.dateTime);
-
-  @override
-  void paint(Canvas canvas, Size size) {
-    double centerX = size.width / 2;
-    double centerY = size.height / 2;
-    Offset center = Offset(centerX, centerY);
-
-    // Hour line
-    canvas.drawLine(
-      center,
-      Offset(280, 150),
-      Paint()
-        ..color = Theme.of(context).colorScheme.secondary
-        ..strokeWidth = 10
-        ..style = PaintingStyle.fill,
-    );
-
-    // Minute line
-    canvas.drawLine(
-      center,
-      Offset(40, 180),
-      Paint()
-        ..color = Theme.of(context).accentColor
-        ..style = PaintingStyle.stroke
-        ..strokeWidth = 10,
-    );
-
-    // Second line
-    canvas.drawLine(center, Offset(80, 150), Paint()..color = Theme.of(context).primaryColor);
-
-    // center Dots
-    Paint dotPainter = Paint()..color = Theme.of(context).primaryIconTheme.color;
-
-    canvas.drawCircle(center, 24, dotPainter);
-    canvas.drawCircle(center, 23, Paint()..color = Theme.of(context).backgroundColor);
-
-    canvas.drawCircle(center, 10, dotPainter);
-  }
-
-  @override
-  bool shouldRepaint(covariant CustomPainter oldDelegate) {
-    return true;
   }
 }
